+++ include/kernel_includes.h
@@ -42,5 +42,6 @@
 #include <fatal.h>
 #include <irq.h>
 #include <sys/thread_stack.h>
+#include <app_memory/mem_domain.h>
 
 #endif /* ZEPHYR_INCLUDE_KERNEL_INCLUDES_H_ */
